Add loading progress bar module
